Move System.Text.Json serializer to Infrastructure project
