+++ src/Umbraco.Cms.Api.Management/DependencyInjection/JsonBuilderExtensions.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Umbraco.Cms.Core.DependencyInjection;
-using Umbraco.Cms.Api.Management.Serialization;
 using Umbraco.Cms.Api.Management.Services;
 
 namespace Umbraco.Cms.Api.Management.DependencyInjection;
@@ -10,8 +9,7 @@ public static class JsonBuilderExtensions
     internal static IUmbracoBuilder AddJson(this IUmbracoBuilder builder)
     {
         builder.Services
-            .AddTransient<IJsonPatchService, JsonPatchService>()
-            .AddTransient<ISystemTextJsonSerializer, SystemTextJsonSerializer>();
+            .AddTransient<IJsonPatchService, JsonPatchService>();
 
         return builder;
     }
