--- src/Umbraco.Cms.Api.Management/Serialization/ISystemTextJsonSerializer.cs
@@ -1,7 +1,0 @@
-﻿using Umbraco.Cms.Core.Serialization;
-
-namespace Umbraco.Cms.Api.Management.Serialization;
-
-public interface ISystemTextJsonSerializer : IJsonSerializer
-{
-}
